Create README.md
The world is suffering from attention deficit. Covit pandemic related data is abundant. After drilling into the data, I've found the results are astonishing. The cumulative affection rate in U.S. is a whopping 8% of its population.  Among the highest infected 20 countries in the world, 17 are in Europe. India is not even among the top 20. Europe is in dire need of vaccines.