US vs. Israel vs. India vs. Canada
Very interesting findings: Israel ranks as the the most vaccinated country with 62 vaccinations per 100 people. But it seems that recent drops in new cases are mostly likely driven by the high level of restriction measures. Maybe it's too early to tell what impacts vaccinations have on the pandemic, given the fact that vaccines take time to be fully effective, not to mention that some vaccines require 2 or more doses over some period.